// src/services/permissions/enums/permission.enum.ts
export enum PermissionEnum {
	VIEW_APPOINTMENTS = 'view appointments',
	CREATE_APPOINTMENTS = 'create appointments',
	EDIT_APPOINTMENTS = 'edit appointments',
	DELETE_APPOINTMENTS = 'delete appointments',

	VIEW_STAFF_SCHEDULES = 'view staff.schedules',
	CREATE_STAFF_SCHEDULES = 'create staff.schedules',
	EDIT_STAFF_SCHEDULES = 'edit staff.schedules',
	DELETE_STAFF_SCHEDULES = 'delete staff.schedules',
}